requirements.txt
ipykernel
numpy
pandas
matplotlib
scikit-learn
seaborn 
scikit-surprise
psutil
flask
kafka-python
confluent-kafka